+++ skorch/tests/test_helper.py
@@ -896,87 +896,3 @@ class TestAccelerate:
         # does not raise
         net.fit(X, y)
         net.predict(X)
-
-
-class TestTrimForPrediction:
-    @pytest.fixture
-    def trim_for_prediction(self):
-        from skorch.helper import trim_for_prediction
-
-        return trim_for_prediction
-
-    @pytest.fixture
-    def net_untrained(self, classifier_module):
-        """A net with a custom 'module2_' criterion and a progress bar callback"""
-        from skorch import NeuralNetClassifier
-        from skorch.callbacks import ProgressBar
-
-        class MyNet(NeuralNetClassifier):
-            def initialize_criterion(self):
-                super().initialize_criterion()
-                # pylint: disable=attribute-defined-outside-init
-                self.criterion2_ = classifier_module()
-                return self
-
-        net = MyNet(
-            classifier_module,
-            max_epochs=2,
-            callbacks=[ProgressBar()],
-        )
-        return net
-
-    @pytest.fixture
-    def net(self, net_untrained, classifier_data):
-        X, y = classifier_data
-        return net_untrained.fit(X[:100], y[:100])
-
-    def test_trim_untrained_net_raises(self, net_untrained, trim_for_prediction):
-        from skorch.exceptions import NotInitializedError
-
-        with pytest.raises(NotInitializedError):
-            trim_for_prediction(net_untrained)
-
-    def test_try_training_trimmed_net(self):
-        # TODO
-        pass
-
-    def test_trim_twice_works(self, net, trim_for_prediction):
-        # does not raise
-        trim_for_prediction(net)
-        trim_for_prediction(net)
-
-    def test_callbacks_trimmed(self, net, trim_for_prediction):
-        trim_for_prediction(net)
-        assert not net.callbacks
-        assert not net.callbacks_
-
-    def test_optimizer_trimmed(self, net, trim_for_prediction):
-        trim_for_prediction(net)
-        assert net.optimizer is None
-        assert net.optimizer_ is None
-
-    def test_criteria_trimmed(self, net, trim_for_prediction):
-        trim_for_prediction(net)
-        assert net.criterion is None
-        assert net.criterion_ is None
-        assert net.criterion2_ is None
-
-    def test_criteria_left_untrimmed(self, net, trim_for_prediction):
-        trim_for_prediction(net, trim_criterion=False)
-        assert net.criterion is not None
-        assert net.criterion_ is not None
-        assert net.criterion2_ is not None
-
-    def test_history_trimmed(self, net, trim_for_prediction):
-        trim_for_prediction(net)
-        assert not net.history
-
-    def test_history_left_untrimmed(self, net, trim_for_prediction):
-        trim_for_prediction(net, trim_history=False)
-        assert net.history
-
-    def test_module_training(self, net, trim_for_prediction):
-        # pylint: disable=protected-access
-        net._set_training(True)
-        trim_for_prediction(net)
-        assert net.module_.training is False
